clean up to new version
