+++ old-src/index.ts
@@ -1,0 +1,3 @@
+export * from './module-manager'
+export * from './plugin-manager'
+export * from './plugins'
