+++ old-src/plugin-manager.ts
