--- src/hacking/api/compiler.api.ts
@@ -1,14 +1,0 @@
-import { ModuleProfile, ModuleApi } from '../remix-module'
-
-export const CompilerProfile: ModuleProfile = {
-  displayName: 'Solidity Compiler',
-  icon: 'compiler',
-  type: 'sol-compiler',
-  methods: [],
-  notifications: []
-}
-
-export class CompilerApi {
-
-
-}
