+++ src/index.ts
@@ -1,3 +1,5 @@
+export * from './iframe.module'
+export * from './internal.module'
 export * from './module-manager'
-export * from './plugin-manager'
-export * from './plugins'
+export * from './remix-module'
+export * from './extension'
